Menambahkan system edit profile

+++ app/Http/Controllers/AccountController.php
@@ -127,21 +127,26 @@ class AccountController extends Controller
      */
     public function edit($nip)
     {
-        if($nip != session()->get('nip')){
-            if(session()->get('nama_jabatan') == "Kepala Sekolah" || session()->get('nama_jabatan') == "Admin"){
+        if($nip == session()->get('nip') || session()->get('nama_jabatan') == "Kepala Sekolah" || session()->get('nama_jabatan') == "Admin"){
+            if($nip != session()->get('nip')){
+                if(session()->get('nama_jabatan') == "Kepala Sekolah" || session()->get('nama_jabatan') == "Admin"){
+                    $akun_data = Akun::find($nip);
+                    $detail_akun = Detailakun::find($akun_data->nip);
+                    $title = "Edit Profil - ";
+                    return view('kepsek.edit-profil',['title' => $title,'akun' => $akun_data, 'detail' => $detail_akun]);
+                }else{
+                    abort(404);
+                }
+            }else{
                 $akun_data = Akun::find($nip);
                 $detail_akun = Detailakun::find($akun_data->nip);
                 $title = "Edit Profil - ";
                 return view('kepsek.edit-profil',['title' => $title,'akun' => $akun_data, 'detail' => $detail_akun]);
-            }else{
-                abort(404);
             }
         }else{
-            $akun_data = Akun::find($nip);
-            $detail_akun = Detailakun::find($akun_data->nip);
-            $title = "Edit Profil - ";
-            return view('kepsek.edit-profil',['title' => $title,'akun' => $akun_data, 'detail' => $detail_akun]);
+            abort(404);
         }
+        
         
     }
 
@@ -152,9 +157,9 @@ class AccountController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request,$id)
+    public function update(Request $request)
     {
-        $this->validate([
+        $this->validate($request,[
             'password' => 'required',
             'nuptk'   => 'required',
             'nama'  => 'required',
@@ -162,38 +167,46 @@ class AccountController extends Controller
             'noHP'  => 'required',
             'alamat' => 'required',
         ]);
-        $akun_data = Akun::find($id);
+        $akun_data = Akun::find($request->nip);
         $detail_akun = Detailakun::find($akun_data->nip);
+
         if($request->hasFile('pic')){
             $avatar = $request->file('pic');
             $filename = time() . "." . $avatar->getClientOriginalExtension();
             Image::make($avatar)->resize(300,300)->save( public_path('img/avatar/' . $filename) );
             
-            
-            if(File::exists(public_path('img/avatar/'.$detail_akun->picture))){
-                File::delete(public_path('img/avatar/'.$detail_akun->picture));
-            }else{
-                //TODO
+            if($detail_akun->picture != 'avatar.jpg'){
+                if(File::exists(public_path('img/avatar/'.$detail_akun->picture))){
+                    File::delete(public_path('img/avatar/'.$detail_akun->picture));
+                }else{
+                    //TODO
+                }
             }
-            
+  
+        }else{
+            $filename = $detail_akun->picture;
         }
-        //TO DO update table detail_akun
+        //TODO update table detail_akun
         $akun_data->password = Hash::make($request->get('password'));
         $detail_akun->nuptk = $request->get('nuptk');
         $detail_akun->nama = $request->get('nama');
         $detail_akun->jk = $request->get('jk');
         $detail_akun->noHP = $request->get('noHP');
+        if($request->get('jurusan')){
+            $detail_akun->id_jurusan = $request->get('jurusan');
+        }
         $detail_akun->alamat = $request->get('alamat');
-        $detail_akun->picture = $request->get('pic');
+        $detail_akun->picture = $filename;
         $akun_data->save();
         $detail_akun->save();
         session([
             'nip' => $akun_data->nip,
-            'nuptk' => $request->get('nuptk'),
-            'nama' => $request->get('nama'),
-            'jk' => $request->get('jk'),
-            'noHP' => $request->get('noHP'),
-            'alamat' => $request->get('alamat'),
+            'ps' => Crypt::encryptString($request->password),
+            'nuptk' => $detail_akun->nuptk,
+            'nama' => $detail_akun->nama,
+            'jk' => $detail_akun->jk,
+            'noHP' => $detail_akun->noHP,
+            'alamat' => $detail_akun->alamat,
             'picture' => $filename
         ]);
         session()->save();
